rename workerfarm to workers
