+++ packages/core/parcel-bundler/src/Bundler.js
@@ -1,7 +1,7 @@
 const fs = require('./utils/fs');
 const Resolver = require('./Resolver');
 const Parser = require('./Parser');
-const WorkerFarm = require('@parcel/workerfarm');
+const WorkerFarm = require('@parcel/workers');
 const Path = require('path');
 const Bundle = require('./Bundle');
 const Watcher = require('./Watcher');
