+++ packages/core/workers/.gitignore
