+++ packages/core/workers/src/.babelrc
