+++ packages/core/workers/src/.eslintrc.json
