+++ packages/core/workers/src/Worker.js
