+++ packages/core/workers/src/child.js
