+++ packages/core/workers/src/cpuCount.js
