+++ packages/core/workers/src/errorUtils.js
