+++ packages/core/workers/test/.babelrc
