+++ packages/core/workers/test/.eslintrc.json
