+++ packages/core/workers/test/integration/workerfarm/echo.js
