+++ packages/core/workers/test/integration/workerfarm/init.js
