+++ packages/core/workers/test/integration/workerfarm/ping.js
