+++ packages/core/workers/test/mocha.opts
